삭제 데이터 파싱 문제 해결

+++ src/main/java/com/handong/cra/crawebbackend/comment/service/CommentServiceImpl.java
@@ -58,9 +58,13 @@ public class CommentServiceImpl implements CommentService {
     @Override
     @Transactional
     public List<ListCommentDto> getCommentsByBoardId(Long boardId) {
+
+        log.info("test here 1");
         Board board = boardRepository.findById(boardId).orElseThrow(BoardNotFoundException::new);
 
+        log.info("test here 2");
         List<Comment> comments = commentRepository.findAllByBoardAndDeletedFalseAndParentCommentIsNull(board);
+        log.info("test here 3");
 
         return comments.stream().map(ListCommentDto::from).toList();
     }
@@ -68,7 +72,7 @@ public class CommentServiceImpl implements CommentService {
     @Override
     @Transactional
     public UpdateCommentDto updateComment(UpdateCommentDto updateCommentDto) {
-        Comment comment = commentRepository.findById(updateCommentDto.getId()).orElseThrow(CommentNotFoundException::new);
+        Comment comment = commentRepository.findWithUserById(updateCommentDto.getId()).orElseThrow(CommentNotFoundException::new);
 
         // 권한 검사
         commentAuthCheck(comment.getUser().getId(), updateCommentDto.getUserId());
@@ -80,16 +84,21 @@ public class CommentServiceImpl implements CommentService {
     @Override
     @Transactional
     public Boolean deleteCommentById(UpdateCommentDto updateCommentDto) {
-        Comment comment = commentRepository.findById(updateCommentDto.getId()).orElseThrow(CommentNotFoundException::new);
+        Comment comment = commentRepository.findWithUserById(updateCommentDto.getId()).orElseThrow(CommentNotFoundException::new);
+
+        if (comment.getDeleted()) // 이미 삭제됨
+            throw new CommentNotFoundException();
 
         // 권한 검사
         commentAuthCheck(comment.getUser().getId(), updateCommentDto.getUserId());
 
         comment.delete();
-
-        List<Comment> comments = comment.getCommentList();
-        // 하위 댓글 삭제 처리
-        for (Comment child : comments) child.delete();
+        // 상위 댓글인 경우
+        if (comment.getParentComment() != null) {
+            List<Comment> comments = comment.getCommentList();
+            // 하위 댓글 삭제 처리
+            for (Comment child : comments) child.delete();
+        }
 
         return true;
     }
